+++ src/Runner.hs
@@ -4,7 +4,7 @@ module Runner
   ( run
   ) where
 
-import Query (Query, Col, selection, condition)
+import Query (Query, Col, selection, condition, evalSel)
 import Element (Row, Elem(..))
 import QueryException (QueryException(..))
 import Data.Map.Ordered as DMO (lookup, fromList)
@@ -13,20 +13,21 @@ import Expression (isTrue)
 
 newtype MissingColumn = MissingColumn Col
 
--- Query -> CSV Rows -> Output Row TODO: This is insane nonsense
+-- Query -> CSV Rows -> Output Row
 run :: Query -> [Row] -> [Row]
 run query = fmap (select query) . filter (whereFilter query)
 
 -- Query Select Columns -> CSV Row -> Selected CSV Row
 select :: Query -> Row -> Row
 select query csvRow =
-  case selectElems selections csvRow of
-    Left (MissingColumn c) -> throw $ MissingColumnsException $ "Table was missing column " <> c
-    Right elems -> DMO.fromList $ zip selections elems
-  where selections = selection query
+  evalSel query csvRow
+--  case selectElems selections csvRow of
+--    Left (MissingColumn c) -> throw $ MissingColumnsException $ "Table was missing column " <> c
+--    Right elems -> DMO.fromList $ zip (evalSel query csvRow) elems
+--  where selections = selection query
 
 selectElems :: [Col] -> Row -> Either MissingColumn [Elem]
-selectElems selections csvRow = mapM (toElem csvRow) selections -- TODO: Combine the MissingColumns
+selectElems selections csvRow = mapM (toElem csvRow) selections
 
 -- CSV Row -> Column name -> Element
 toElem :: Row -> Col -> Either MissingColumn Elem