+++ test/fixtures/queries/select_and_where_type_fail.sql
@@ -1,0 +1,3 @@
+SELECT NOT age AS notAge
+FROM test/fixtures/tables/people
+WHERE NOT age