+++ test/fixtures/queries/select_as.sql
@@ -1,0 +1,2 @@
+SELECT first_name AS name, 3 as three
+FROM test/fixtures/tables/people