+++ test/fixtures/queries/select_type_fail_and.sql
@@ -1,0 +1,2 @@
+SELECT age AND True AS ageTrue
+FROM test/fixtures/tables/people