+++ test/fixtures/queries/select_type_fail_no_as.sql
@@ -1,0 +1,2 @@
+SELECT 2 + 3
+FROM test/fixtures/tables/people