+++ test/fixtures/queries/selection_expr_dup_cols.sql
@@ -1,0 +1,3 @@
+SELECT first_name ++ " B" ++ first_name ++ " " ++ last_name aS long_name, age * age * 2 AS overbig_age
+FROM test/fixtures/tables/people
+WHERE last_name = "McKnight"