+++ test/fixtures/queries/where_type_fail_cat.sql
@@ -1,3 +1,3 @@
 SELECT age
 FROM test/fixtures/tables/people_camel
-WHERE "Hello #" ++ 3
+WHERE age ++ 3