+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # Changelog for csvsql
 
+## v0.0.7
+* SELECT Expressions
+* SELECT AS
+* SELECT Expression type checking
+* SELECT * support
+
 ### v0.0.6
 * Type Checking of WHERE clause